Change the python-devel BuildRequires to python-libs
We want a BuildRequires for the distutils module, since it's not in
the regular python package (surprise!), but python-libs is where it
actually lives. python-devel is more for Python/C stuff, and we don't
have any of that.

(cherry picked from commit 3150445a814d2e5598ddf2a0adbd41a7ca8cf6a7)

Conflicts:
	fedup.spec

Related: rhbz#1012617