Change the python-devel BuildRequires to python-libs
We want a BuildRequires for the distutils module, since it's not in
the regular python package (surprise!), but python-libs is where it
actually lives. python-devel is more for Python/C stuff, and we don't
have any of that.

(cherry picked from commit 4e1faf6713a49cc6fb0967be59ecd61b4a8bd9a9)

Related: rhbz#1084143