Fix ValueError with --addrepo/--repo REPOID
We need to be sure that the argument to --addrepo is actually of the
form "REPOID=[@]URL", or we end up with a ValueError when we try to
split it into two parts later on. So let's check that early.

Also, the --repo argument was deprecated and hidden, and it's now
causing confusion with --instrepo, so drop it.

(cherry picked from commit 471b7c7207a279093449eb08d1ef8b4fc59ee0e1)

Related: rhbz#1085023