Treeinfo file renamed product section (RHEL 6)
Format of the .treeinfo file has been changed in RHEL 7.4 repositories. RHEL repositories prior to 7.4 version had .treeinfo containing 'product' section which was not according to the specification [1]. To correct that, it was changed to 'release' section. Hence, redhat-upgrade-tool has been updated to allow both sections, 'product' as a default and 'release' as a fallback (thanks Marian Ganisin for the patch).

Related: RHBZ#[1456809](https://bugzilla.redhat.com/show_bug.cgi?id=1456809)

[1] http://productmd.readthedocs.io/en/latest/treeinfo-1.1.html