Remove files of the target kernel under /boot
After the rollback, kernel / bootloader files of target kernel have
been kept in the /boot parition. As easier solution of this,
I put cleaning of this file automatically when the --clean option is
used. As we require run of r-u-t with the --clean option after the
rollback, it is for now safe solution.

To ensure that we remove only files that are created during the
upgrade and will not remove kernel files that it is available
under dual boot (unexpected, but just to be sure), it is checked
version of target kernel and in case `kernelpath` of this version
doesn't exist already, information is stored under `rollback_dir`
in the `.target-kernel` file. Only in case the data is stored in the
file, kernel files of target kernel will be cleaned.