Fix chmod and rollback import
When os.chmod is used, use of 0xxx mode is incorrect. Either xxx
is wrong as mode is not set as expected. Only correct way here is
use of 0oxxx format. E.g.:
    0o775

Additionaly, current solution around rollbacks directly from RHEL 7
system really needs special run and it requires different import
when it is used on RHEL 6 and when it is used on RHEL 7 to provide
a rollback or cleanup.