Bump version to 0.8.0
As the rollbacks are the most probably the largest change we did in
the last years in this tool and probably there will not be another
so big change in future, we decided to bump the minor version as
this release contains a lot of changes - even when original
fuctionality should be untouched by these changes in case user will
not use the rollback capability for upgrades.

Praise the sun!